refactoring and add routing
