+++ project/public/index.html
@@ -7,7 +7,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>WTW</title>
   <meta name="robots" content="noindex, nofollow">
-  <link rel="stylesheet" href="project/public/css/main.min.css">
+  <link rel="stylesheet" href="css/main.min.css">
 </head>
 
 <body>
