+++ project/src/components/main/catalog-genres/catalog-genres-data.tsx
@@ -1,0 +1,2 @@
+export const catalogGenresData : string[] = ['Comedies', 'Crime', 'Documentary', 'Dramas', 'Horror', 'Kids & Family',
+  'Romance', 'Sci-Fi', 'Thrillers'];
