+++ project/src/components/main/film-card/film-card.tsx
