+++ project/src/components/sign-in/error.tsx
@@ -1,0 +1,9 @@
+function Error(): JSX.Element {
+  return (
+    <div className="sign-in__message">
+      <p>Please enter a valid email address</p>
+    </div>
+  );
+}
+
+export default Error;
