+++ project/src/components/sign-in/message.tsx
@@ -1,0 +1,9 @@
+function Message(): JSX.Element {
+  return (
+    <div className="sign-in__message">
+      <p>We can’t recognize this email <br/> and password combination. Please try again.</p>
+    </div>
+  );
+}
+
+export default Message;
