+++ project/src/pages/add-review-screen/add-review-screen.tsx
@@ -1,147 +1,93 @@
+import Logo from '../../components/logo/logo';
+
 function AddReviewScreen(): JSX.Element {
   return (
-    <body>
-      <div className="visually-hidden">
-        <svg xmlns="http://www.w3.org/2000/svg" xmlnsXlink="http://www.w3.org/1999/xlink">
-          <symbol id="add" viewBox="0 0 19 20">
-            <title>+</title>
-            <desc>Created with Sketch.</desc>
-            <g id="Page-1" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-              <polygon id="+" fill="#EEE5B5" points="10.777832 11.2880859 10.777832 19.5527344 8.41650391 19.5527344 8.41650391 11.2880859 0.627929688 11.2880859 0.627929688 8.92675781 8.41650391 8.92675781 8.41650391 0.662109375 10.777832 0.662109375 10.777832 8.92675781 18.5664062 8.92675781 18.5664062 11.2880859"/>
-            </g>
-          </symbol>
-          <symbol id="full-screen" viewBox="0 0 27 27">
-            <path fillRule="evenodd" clipRule="evenodd" d="M23.8571 0H16V3.14286H23.8571V11H27V3.14286V0H23.8571Z"
-              fill="#FFF9D9" fillOpacity="0.7"
-            />
-            <path fillRule="evenodd" clipRule="evenodd" d="M27 23.8571V16H23.8571V23.8571H16V27H23.8571H27L27 23.8571Z"
-              fill="#FFF9D9" fillOpacity="0.7"
-            />
-            <path fillRule="evenodd" clipRule="evenodd"
-              d="M0 3.14286L0 11H3.14286L3.14286 3.14286L11 3.14286V0H3.14286H0L0 3.14286Z" fill="#FFF9D9"
-              fillOpacity="0.7"
-            />
-            <path fillRule="evenodd" clipRule="evenodd"
-              d="M3.14286 27H11V23.8571H3.14286L3.14286 16H0L0 23.8571V27H3.14286Z" fill="#FFF9D9"
-              fillOpacity="0.7"
-            />
-          </symbol>
-          <symbol id="in-list" viewBox="0 0 18 14">
-            <path fillRule="evenodd" clipRule="evenodd"
-              d="M2.40513 5.35353L6.1818 8.90902L15.5807 0L18 2.80485L6.18935 14L0 8.17346L2.40513 5.35353Z"
-              fill="#EEE5B5"
-            />
-          </symbol>
-          <symbol id="pause" viewBox="0 0 14 21">
-            <title>Artboard</title>
-            <desc>Created with Sketch.</desc>
-            <g id="Artboard" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-              <polygon id="Line" fill="#EEE5B5" fillRule="nonzero"
-                points="0 -1.11910481e-13 4 -1.11910481e-13 4 21 0 21"
-              />
-              <polygon id="Line" fill="#EEE5B5" fillRule="nonzero"
-                points="10 -1.11910481e-13 14 -1.11910481e-13 14 21 10 21"
-              />
-            </g>
-          </symbol>
-        </svg>
-      </div>
+    <section className="film-card film-card--full">
+      <div className="film-card__header">
+        <div className="film-card__bg">
+          <img src="img/bg-the-grand-budapest-hotel.jpg" alt="The Grand Budapest Hotel"/>
+        </div>
 
-      <section className="film-card film-card--full">
-        <div className="film-card__header">
-          <div className="film-card__bg">
-            <img src="img/bg-the-grand-budapest-hotel.jpg" alt="The Grand Budapest Hotel"/>
-          </div>
+        <h1 className="visually-hidden">WTW</h1>
 
-          <h1 className="visually-hidden">WTW</h1>
-
-          <header className="page-header">
-            <div className="logo">
-              <a href="main.html" className="logo__link">
-                <span className="logo__letter logo__letter--1">W</span>
-                <span className="logo__letter logo__letter--2">T</span>
-                <span className="logo__letter logo__letter--3">W</span>
-              </a>
-            </div>
-
-            <nav className="breadcrumbs">
-              <ul className="breadcrumbs__list">
-                <li className="breadcrumbs__item">
-                  <a href="film-page.html" className="breadcrumbs__link">The Grand Budapest Hotel</a>
-                </li>
-                <li className="breadcrumbs__item">
-                  <a className="breadcrumbs__link">Add review</a>
-                </li>
-              </ul>
-            </nav>
-
-            <ul className="user-block">
-              <li className="user-block__item">
-                <div className="user-block__avatar">
-                  <img src="img/avatar.jpg" alt="User avatar" width="63" height="63"/>
-                </div>
+        <header className="page-header">
+          <Logo/>
+          <nav className="breadcrumbs">
+            <ul className="breadcrumbs__list">
+              <li className="breadcrumbs__item">
+                <a href="film-page.html" className="breadcrumbs__link">The Grand Budapest Hotel</a>
               </li>
-              <li className="user-block__item">
-                <a className="user-block__link">Sign out</a>
+              <li className="breadcrumbs__item">
+                <a className="breadcrumbs__link">Add review</a>
               </li>
             </ul>
-          </header>
+          </nav>
 
-          <div className="film-card__poster film-card__poster--small">
-            <img src="img/the-grand-budapest-hotel-poster.jpg" alt="The Grand Budapest Hotel poster"
-              width="218" height="327"
-            />
+          <ul className="user-block">
+            <li className="user-block__item">
+              <div className="user-block__avatar">
+                <img src="img/avatar.jpg" alt="User avatar" width="63" height="63"/>
+              </div>
+            </li>
+            <li className="user-block__item">
+              <a className="user-block__link">Sign out</a>
+            </li>
+          </ul>
+        </header>
+
+        <div className="film-card__poster film-card__poster--small">
+          <img src="img/the-grand-budapest-hotel-poster.jpg" alt="The Grand Budapest Hotel poster"
+            width="218" height="327"
+          />
+        </div>
+      </div>
+
+      <div className="add-review">
+        <form action="#" className="add-review__form">
+          <div className="rating">
+            <div className="rating__stars">
+              <input className="rating__input" id="star-10" type="radio" name="rating" value="10"/>
+              <label className="rating__label" htmlFor="star-10">Rating 10</label>
+
+              <input className="rating__input" id="star-9" type="radio" name="rating" value="9"/>
+              <label className="rating__label" htmlFor="star-9">Rating 9</label>
+
+              <input className="rating__input" id="star-8" type="radio" name="rating" value="8" checked/>
+              <label className="rating__label" htmlFor="star-8">Rating 8</label>
+
+              <input className="rating__input" id="star-7" type="radio" name="rating" value="7"/>
+              <label className="rating__label" htmlFor="star-7">Rating 7</label>
+
+              <input className="rating__input" id="star-6" type="radio" name="rating" value="6"/>
+              <label className="rating__label" htmlFor="star-6">Rating 6</label>
+
+              <input className="rating__input" id="star-5" type="radio" name="rating" value="5"/>
+              <label className="rating__label" htmlFor="star-5">Rating 5</label>
+
+              <input className="rating__input" id="star-4" type="radio" name="rating" value="4"/>
+              <label className="rating__label" htmlFor="star-4">Rating 4</label>
+
+              <input className="rating__input" id="star-3" type="radio" name="rating" value="3"/>
+              <label className="rating__label" htmlFor="star-3">Rating 3</label>
+
+              <input className="rating__input" id="star-2" type="radio" name="rating" value="2"/>
+              <label className="rating__label" htmlFor="star-2">Rating 2</label>
+
+              <input className="rating__input" id="star-1" type="radio" name="rating" value="1"/>
+              <label className="rating__label" htmlFor="star-1">Rating 1</label>
+            </div>
           </div>
-        </div>
 
-        <div className="add-review">
-          <form action="#" className="add-review__form">
-            <div className="rating">
-              <div className="rating__stars">
-                <input className="rating__input" id="star-10" type="radio" name="rating" value="10"/>
-                <label className="rating__label" htmlFor="star-10">Rating 10</label>
-
-                <input className="rating__input" id="star-9" type="radio" name="rating" value="9"/>
-                <label className="rating__label" htmlFor="star-9">Rating 9</label>
-
-                <input className="rating__input" id="star-8" type="radio" name="rating" value="8" checked/>
-                <label className="rating__label" htmlFor="star-8">Rating 8</label>
-
-                <input className="rating__input" id="star-7" type="radio" name="rating" value="7"/>
-                <label className="rating__label" htmlFor="star-7">Rating 7</label>
-
-                <input className="rating__input" id="star-6" type="radio" name="rating" value="6"/>
-                <label className="rating__label" htmlFor="star-6">Rating 6</label>
-
-                <input className="rating__input" id="star-5" type="radio" name="rating" value="5"/>
-                <label className="rating__label" htmlFor="star-5">Rating 5</label>
-
-                <input className="rating__input" id="star-4" type="radio" name="rating" value="4"/>
-                <label className="rating__label" htmlFor="star-4">Rating 4</label>
-
-                <input className="rating__input" id="star-3" type="radio" name="rating" value="3"/>
-                <label className="rating__label" htmlFor="star-3">Rating 3</label>
-
-                <input className="rating__input" id="star-2" type="radio" name="rating" value="2"/>
-                <label className="rating__label" htmlFor="star-2">Rating 2</label>
-
-                <input className="rating__input" id="star-1" type="radio" name="rating" value="1"/>
-                <label className="rating__label" htmlFor="star-1">Rating 1</label>
-              </div>
+          <div className="add-review__text">
+            <textarea className="add-review__textarea" name="review-text" id="review-text" placeholder="Review text"/>
+            <div className="add-review__submit">
+              <button className="add-review__btn" type="submit">Post</button>
             </div>
 
-            <div className="add-review__text">
-              <textarea className="add-review__textarea" name="review-text" id="review-text" placeholder="Review text"/>
-              <div className="add-review__submit">
-                <button className="add-review__btn" type="submit">Post</button>
-              </div>
-
-            </div>
-          </form>
-        </div>
-
-      </section>
-    </body>
+          </div>
+        </form>
+      </div>
+    </section>
   );
 }
