+++ project/src/pages/main-screen/main-screen.tsx
@@ -1,5 +1,9 @@
-import FilmCard from '../../components/film-card/film-card';
-import {filmCards} from '../../film-cards-data';
+import FilmCard from '../../components/main/film-card/film-card';
+import {filmCards} from '../../components/main/film-card/film-cards-data';
+import Logo from '../../components/logo/logo';
+import Footer from '../../components/footer/footer';
+import {catalogGenresData} from '../../components/main/catalog-genres/catalog-genres-data';
+import CatalogGenres from '../../components/main/catalog-genres/catalog-genres';
 
 type MainScreenProps = {
   title: string;
@@ -18,13 +22,7 @@ function MainScreen({title, genre, year} : MainScreenProps): JSX.Element {
         <h1 className="visually-hidden">WTW</h1>
 
         <header className="page-header film-card__head">
-          <div className="logo">
-            <a className="logo__link">
-              <span className="logo__letter logo__letter--1">W</span>
-              <span className="logo__letter logo__letter--2">T</span>
-              <span className="logo__letter logo__letter--3">W</span>
-            </a>
-          </div>
+          <Logo/>
 
           <ul className="user-block">
             <li className="user-block__item">
@@ -81,33 +79,7 @@ function MainScreen({title, genre, year} : MainScreenProps): JSX.Element {
             <li className="catalog__genres-item catalog__genres-item--active">
               <a href="#" className="catalog__genres-link">All genres</a>
             </li>
-            <li className="catalog__genres-item">
-              <a href="#" className="catalog__genres-link">Comedies</a>
-            </li>
-            <li className="catalog__genres-item">
-              <a href="#" className="catalog__genres-link">Crime</a>
-            </li>
-            <li className="catalog__genres-item">
-              <a href="#" className="catalog__genres-link">Documentary</a>
-            </li>
-            <li className="catalog__genres-item">
-              <a href="#" className="catalog__genres-link">Dramas</a>
-            </li>
-            <li className="catalog__genres-item">
-              <a href="#" className="catalog__genres-link">Horror</a>
-            </li>
-            <li className="catalog__genres-item">
-              <a href="#" className="catalog__genres-link">Kids & Family</a>
-            </li>
-            <li className="catalog__genres-item">
-              <a href="#" className="catalog__genres-link">Romance</a>
-            </li>
-            <li className="catalog__genres-item">
-              <a href="#" className="catalog__genres-link">Sci-Fi</a>
-            </li>
-            <li className="catalog__genres-item">
-              <a href="#" className="catalog__genres-link">Thrillers</a>
-            </li>
+            {catalogGenresData.map((item) => (<CatalogGenres genre={item}/>))}
           </ul>
 
           <div className="catalog__films-list">
@@ -117,20 +89,7 @@ function MainScreen({title, genre, year} : MainScreenProps): JSX.Element {
             <button className="catalog__button" type="button">Show more</button>
           </div>
         </section>
-
-        <footer className="page-footer">
-          <div className="logo">
-            <a className="logo__link logo__link--light">
-              <span className="logo__letter logo__letter--1">W</span>
-              <span className="logo__letter logo__letter--2">T</span>
-              <span className="logo__letter logo__letter--3">W</span>
-            </a>
-          </div>
-
-          <div className="copyright">
-            <p>© 2019 What to watch Ltd.</p>
-          </div>
-        </footer>
+        <Footer/>
       </div>
     </>
   );
