+++ project/src/pages/not-found/not-found.tsx
@@ -1,0 +1,15 @@
+import {Link} from 'react-router-dom';
+
+function NotFound(): JSX.Element {
+  return (
+    <>
+      <h1>404.
+        <br/>
+        <small>Page not found</small>
+      </h1>
+      <Link to="/">Go to main page </Link>
+    </>
+  );
+}
+
+export default NotFound;
